engine/vinylcontrolcontrol: Fix precision loss (wrong use of int)

+++ src/engine/vinylcontrolcontrol.cpp
@@ -117,7 +117,7 @@ void VinylControlControl::slotControlVinylSeek(double fractionalPos) {
         }
 
         double shortest_distance = 0;
-        int nearest_playpos = -1;
+        double nearest_playpos = -1;
 
         const QList<CuePointer> cuePoints(pTrack->getCuePoints());
         QListIterator<CuePointer> it(cuePoints);
@@ -127,7 +127,7 @@ void VinylControlControl::slotControlVinylSeek(double fractionalPos) {
                 continue;
             }
 
-            int cue_position = pCue->getPosition();
+            double cue_position = pCue->getPosition();
             // pick cues closest to new_playpos
             if ((nearest_playpos == -1) ||
                 (fabs(new_playpos - cue_position) < shortest_distance)) {
